Int health system converted to handle floats.

+++ Turn-based combat/Assets/Scripts/Unit.cs
@@ -8,12 +8,12 @@ public class Unit : MonoBehaviour
 	public string unitName;
 	public int unitLevel;
 
-	public int damage;
+	public float damage;
 
-	public int maxHP;
-	public int currentHP;
+	public float maxHP;
+	public float currentHP;
 
-	public bool TakeDamage(int dmg)
+	public bool TakeDamage(float dmg)
 	{
 		currentHP -= dmg;
 
@@ -23,7 +23,7 @@ public class Unit : MonoBehaviour
 			return false;
 	}
 
-	public void Heal(int amount)
+	public void Heal(float amount)
 	{
 		currentHP += amount;
 		if (currentHP > maxHP)
